Simplify Salamander implementation
