+++ src/salamatrix-sdk/native-ui-runtime/salamatrix_ui_win32_host.h
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
+#pragma once
+#include "salamatrix_ui_host.h"
+
+namespace Salamatrix { namespace UI {
+INativeDialogHost* WINAPI GetWin32NativeDialogHost();
+} }
