+++ src/tools/salamatrix-studio/CHANGELOG.md
@@ -8,3 +8,5 @@
 - PowerShell, Python, Node.js, PHP, Lua, Automation JScript, and native C++ dialog generation.
 - Visual extension manifest and menu designer.
 - Bundled standalone x64 native preview host and versioned SDK boundary.
+- Shared NativeDialog and Salamatrix-specific controls used identically by
+  Salamatrix.SPL and the Studio preview host.
