+++ src/tools/salamatrix-studio/test/fixtures/basic.smxp
@@ -1,3 +1,8 @@
 SMXPREVIEW1
 50726576696577
 320	180
+636f6d626f626f78	63686f696365		8	8	120	80	0	0	0
+O	726561644f6e6c79	boolean	66616c7365
+I	4f6e65
+S	0
+V	1	5265717569726564
